Window management update (#17)
* Split window and resizable composables further
