+++ src/components/AppResizableContainer/useResizableStyle.ts
@@ -1,0 +1,67 @@
+/*
+* This file is part of daw.ts.
+*
+* Copyright (c) 2024 Yann Mazita
+*
+* Daw.ts is free software: you can redistribute it and/or modify
+* it under the terms of the GNU General Public License as published by
+* the Free Software Foundation, either version 3 of the License, or
+* any later version.
+*
+* Daw.ts is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+* GNU General Public License for more details.
+*
+* You should have received a copy of the GNU General Public License
+* along with daw.ts. If not, see https://www.gnu.org/licenses/.
+*/
+
+/*
+* Original code derived from vue-resizable,
+* licensed under the MIT License.
+*
+* Copyright (c) 2018 nikitasnv
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in all
+* copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+* SOFTWARE.
+*/
+
+
+import { computed } from 'vue';
+import { CALC_MASK } from './constants';
+import { useResizableState } from './useResizableState';
+
+export function useResizableStyle(state: ReturnType<typeof useResizableState>) {
+    const { w, h, l, t, calcMap } = state;
+
+    return computed(() => ({
+        ...(calcMap.value & CALC_MASK.w && {
+            width: typeof w.value === "number" ? `${w.value}px` : w.value,
+        }),
+        ...(calcMap.value & CALC_MASK.h && {
+            height: typeof h.value === "number" ? `${h.value}px` : h.value,
+        }),
+        ...(calcMap.value & CALC_MASK.l && {
+            left: typeof l.value === "number" ? `${l.value}px` : l.value,
+        }),
+        ...(calcMap.value & CALC_MASK.t && {
+            top: typeof t.value === "number" ? `${t.value}px` : t.value,
+        }),
+    }));
+}
